:truck: Move alert component to modal folder
