+++ src/kirby/components/modal/alert/alert.component.scss
@@ -1,4 +1,4 @@
-@import "../../scss/utils";
+@import "../../../scss/utils";
 
 h3.title {
   margin-top: size('m');
