+++ src/kirby/kirby.module.ts
@@ -12,7 +12,7 @@ import { KeyHandlerDirective } from '@kirbydesign/designsystem/directives/key-ha
 import { FullscreenLoadingOverlayComponent } from './components/loading-overlay/fullscreen-loading-overlay/fullscreen-loading-overlay.component';
 import { LoadingOverlayComponent } from './components/loading-overlay/loading-overlay.component';
 import { ListItemColorDirective } from './components/list/directives/list-item-color.directive';
-import { AlertComponent } from './components/alert/alert.component';
+import { AlertComponent } from './components/modal/alert/alert.component';
 import { PageModule } from '@kirbydesign/designsystem/components/page/page.module';
 import { FormFieldMessageComponent } from './components/form-field/form-field-message/form-field-message.component';
 import { AppModule } from '@kirbydesign/designsystem/components/app/app.module';
